+++ src/main/java/net/rptools/maptool/client/ui/zone/renderer/LabelLocation.java
@@ -1,0 +1,29 @@
+/*
+ * This software Copyright by the RPTools.net development team, and
+ * licensed under the Affero GPL Version 3 or, at your option, any later
+ * version.
+ *
+ * MapTool Source Code is distributed in the hope that it will be
+ * useful, but WITHOUT ANY WARRANTY; without even the implied warranty
+ * of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ *
+ * You should have received a copy of the GNU Affero General Public
+ * License * along with this source Code.  If not, please visit
+ * <http://www.gnu.org/licenses/> and specifically the Affero license
+ * text at <http://www.gnu.org/licenses/agpl.html>.
+ */
+package net.rptools.maptool.client.ui.zone.renderer;
+
+import java.awt.*;
+import net.rptools.maptool.model.Label;
+
+class LabelLocation {
+
+  public Rectangle bounds;
+  public net.rptools.maptool.model.Label label;
+
+  public LabelLocation(Rectangle bounds, Label label) {
+    this.bounds = bounds;
+    this.label = label;
+  }
+}
