+++ src/main/java/net/rptools/maptool/model/Path.java
@@ -18,8 +18,8 @@ import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
 import net.rptools.maptool.client.MapTool;
+import net.rptools.maptool.client.ui.zone.renderer.SelectionSet;
 import net.rptools.maptool.client.ui.zone.renderer.ZoneRenderer;
-import net.rptools.maptool.client.ui.zone.renderer.ZoneRenderer.SelectionSet;
 import net.rptools.maptool.client.walker.NaiveWalker;
 import net.rptools.maptool.server.proto.PathDto;
 import net.rptools.maptool.server.proto.drawing.IntPointDto;
